update chức năng nhà cung cấp-module mua hàng - cần thêm builder.Services.AddScoped<SupplierRepo>(); vào program
Found a swap file by the name "/d/Duan_TeamA_Project/RTC_Web_Angular_V2/RERP_BACKEND_V2/.git/.COMMIT_EDITMSG.swp"
          owned by: NEM   dated: Fri Nov 21 09:15:33 2025
         file name: /d/Duan_TeamA_Project/RTC_Web_Angular_V2/RERP_BACKEND_V2/.git/COMMIT_EDITMSG
          modified: YES
         user name: NEM   host name: DESKTOP-HK5JPET
        process ID: 251
-- More --
