+++ RERPAPI.Repo/GenericEntity/SupplierRepo.cs
@@ -1,0 +1,17 @@
+﻿using RERPAPI.Model.DTO;
+using RERPAPI.Model.Entities;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RERPAPI.Repo.GenericEntity
+{
+    public class SupplierRepo : GenericRepo<Supplier>
+    {
+        public SupplierRepo(CurrentUser currentUser) : base(currentUser)
+        {
+        }
+    }
+}
